Fix for #55
Fixes #55
New button `Finish` in shopping list page. Clicking this button updates
the product list. The number bought from the shoppoing list will be
added to the amount of each product the is marked as `in cart` (by
clicking on the product name or the check mark).
There will be an api call for each product as no batch api call is
available. After the call to the api to change the amount, the number
bought is set to 0, so that multiple clicks on the `Finish` button will
not change the amount of the product further.

Maybe there should be some feedback, that the values will only be
updated when the product is marked as `in cart`. Simply toggling the
enabled status of the `Finish` button would help.